style: MyTest를 각 디렉토리 별 테스트로 분리/저장
