+++ __tests__/BridgeGameTest.js
@@ -1,94 +1,4 @@
-const { ERROR } = require("../src/data/Constants");
-const Validation = require("../src/Validation");
-
-const BridgeMaker = require("../src/BridgeMaker");
-const BridgeRandomNumberGenerator = require("../src/BridgeRandomNumberGenerator");
 const BridgeGame = require("../src/BridgeGame");
-const { checkBridgeLength } = require("../src/Validation");
-
-describe(`bridgeSize 입력값 타당성 테스트`, () => {
-  test.each([["0"], ["-1"], ["21"]])(
-    `유저가 3이상 20이하의 숫자를 입력했는지 확인`,
-    (input) => {
-      expect(() => checkBridgeLength(input)).toThrow(
-        ERROR.ERROR_BRIDGE_LENGTH_RANGE
-      );
-    }
-  );
-
-  test.each([["a"], ["a1"], ["#3"], ["10."]])(
-    `유저가 숫자만 입력 했는지 확인`,
-    (input) => {
-      expect(() => {
-        Validation.checkBridgeLength(input);
-      }).toThrow(ERROR.ERROR_BRIDGE_LENGTH_ONLY_NUM);
-    }
-  );
-});
-
-describe(`다리를 생성기능 테스트`, () => {
-  test(`입력받은 bridgeSize와 만들어진 다리 배열의 길이는 일치한다`, () => {
-    const size = 3;
-    const generateRandomNumber = BridgeRandomNumberGenerator.generate;
-
-    expect(BridgeMaker.makeBridge(size, generateRandomNumber)).toHaveLength(
-      size
-    );
-  });
-
-  test(`생성된 랜덤 숫자가 1이면 "U"를 배열에 저장한다.`, () => {
-    const size = 3;
-    function generateRandomNumber() {
-      return 1;
-    }
-    expect(BridgeMaker.makeBridge(size, generateRandomNumber)).toEqual([
-      "U",
-      "U",
-      "U",
-    ]);
-  });
-
-  test(`생성된 랜덤 숫자가 0이면 "D"를 배열에 저장한다`, () => {
-    const size = 3;
-    function generateRandomNumber() {
-      return 0;
-    }
-    expect(BridgeMaker.makeBridge(size, generateRandomNumber)).toEqual([
-      "D",
-      "D",
-      "D",
-    ]);
-  });
-
-  test(`생선된 랜덤 숫자는 0 또는 1 이다.`, () => {
-    expect(String(BridgeRandomNumberGenerator.generate())).toMatch(/0|1/);
-  });
-});
-
-describe(`userMove 입력값 타당성 테스트`, () => {
-  test.each([["u"], ["d"]])(
-    `알파벳이 UpperCase가 아닐 때 에러발생`,
-    (input) => {
-      expect(() => {
-        Validation.checkMove(input);
-      }).toThrow(ERROR.ERROR_BRIDGE_MOVE_UPPERCASE);
-    }
-  );
-
-  test.each([["#d"], ["a"], ["1d"], ["123"], ["U%"], ["D#"]])(
-    `U 또는 D 가 아닌 다른 걸 입력했을 때 에러발생`,
-    (input) => {
-      expect(() => {
-        Validation.checkMove(input);
-      }).toThrow(ERROR.ERROR_BRIDGE_MOVE_RANGE);
-    }
-  );
-  test.each([["DD"], ["UU"]])(`2글자 이상을 입력했을 때 에러발생`, (input) => {
-    expect(() => {
-      Validation.checkMove(input);
-    }).toThrow(ERROR.ERROR_BRIDGE_MOVE_LENGTH);
-  });
-});
 
 describe(`유저 이동 결과 판단 및 결과를 배열에 저장 테스트`, () => {
   test(`"U"을 입력했을 때 유저가 다리 건너기에 <실패>하면 이중 배열에 그 결과값을 저장한다.`, () => {
@@ -174,35 +84,6 @@ describe(`유저 이동 결과 판단 및 결과를 배열에 저장 테스트`,
     bridgeGame.move(bridge, userMove);
     expect(bridgeGame.getResult()).toEqual([["O"], [` `]]);
   });
-});
-
-describe(`재시작 입력값 타당성 테스트`, () => {
-  test.each([["r"], ["q"]])(
-    `알파벳이 UpperCase가 아닐 때 에러발생`,
-    (input) => {
-      expect(() => {
-        Validation.checkRetry(input);
-      }).toThrow(ERROR.ERROR_RETRY_UPPERCASE);
-    }
-  );
-
-  test.each([["a"], ["#"], ["Q."], ["R3"]])(
-    `R 또는 Q가 아닌 다른 걸 입력했을 때 에러발생`,
-    (input) => {
-      expect(() => {
-        Validation.checkRetry(input);
-      }).toThrow(ERROR.ERROR_RETRY_RANGE);
-    }
-  );
-
-  test.each([["RR"], ["QQ"], ["QR"], ["RQ"]])(
-    `2글자 이상을 입력했을 때 에러발생`,
-    (input) => {
-      expect(() => {
-        Validation.checkRetry(input);
-      }).toThrow(ERROR.ERROR_RETRY_LENGTH);
-    }
-  );
 });
 
 describe(`입력값에 따라 재시작/종료 판단`, () => {
